data/data/com.termux/files/home/CompassAIPlatform/frontend/backend_temp/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
motor==3.3.2
pymongo==4.7.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
websockets==12.0
grpcio==1.64.0
grpcio-tools==1.64.0
langchain==0.2.5
